small optimizations

+++ cpp/util/display.cpp
@@ -3,13 +3,13 @@
 #include "display.hpp"
 
 Progress::Progress(size_t max)
-    : iteration{ 0 }, max{ max }
+    : iteration{ 0 }, max{ (uint32_t)max }
 {}
 
 void Progress::update()
 {
     iteration++;
-    std::printf("\rIteration [%d / %d]", (int)iteration, (int)max);
+    std::printf("\rIteration [%d / %d]", iteration, max);
     fflush(stdout);
     if (iteration == max)
         std::cout << std::endl;
