+++ cpp/util/display.hpp
@@ -5,8 +5,8 @@
 
 struct Progress {
     Progress(size_t max);
-    size_t iteration = 0;
-    size_t max = 0;
+    uint32_t iteration = 0;
+    uint32_t max = 0;
 
     void update();
 };
